+++ crates/fixed-point-wasm/src/lib.rs
@@ -1,8 +1,10 @@
+mod utils;
+
 use core::fmt;
 use std::str::FromStr;
 
 use delv_core::{
-    conversions::{ToBigInt, ToFixedPoint, ToI256},
+    conversions::{ToBigInt, ToFixedPoint},
     error,
     error::{Error, ToResult},
 };
@@ -11,6 +13,7 @@ use fixedpointmath::{uint256, FixedPoint};
 use js_sys::{parse_float, BigInt, JsString};
 use rand::{thread_rng, Rng};
 use ts_macro::ts;
+pub use utils::*;
 use wasm_bindgen::prelude::*;
 
 // Initialization function
@@ -266,81 +269,7 @@ impl Fixed {
     }
 }
 
-// Utils //
-
-/// Create a new fixed-point number from a scaled value.
-///
-/// @param value - A scaled value.
-///
-/// @param decimals - The number of decimal places to use. Max is `18`.
-/// Defaults to `18`.
-///
-/// @example
-//
-/// ```js
-/// const fromBigint = fixed(1500000000000000000n);
-/// const fromNumber = fixed(1.5e18);
-/// const fromString = fixed('1.5e18');
-/// const withDecimals = fixed(1.5e6, 6);
-///
-/// console.log(fromBigint.toString());
-/// // => 1.500000000000000000
-///
-/// console.log(fromNumber.toString());
-/// // => 1.500000000000000000
-///
-/// console.log(fromString.toString());
-/// // => 1.500000000000000000
-///
-/// console.log(withDecimals.toString());
-/// // => 1.500000
-/// ```
-#[wasm_bindgen(skip_jsdoc)]
-pub fn fixed(value: Numberish, decimals: Option<u8>) -> Result<Fixed, Error> {
-    Fixed::new(Some(value), decimals)
-}
-
-/// Create a random fixed-point number with an optional min and max.
-///
-/// @example
-///
-/// ```ts
-/// const random = randomFixed();
-///
-/// console.log(random.toString());
-/// // => 0.472987274007185487
-/// ```
-#[wasm_bindgen(skip_jsdoc, js_name = randomFixed)]
-pub fn random_fixed(params: Option<IGenerateRandomParams>) -> Result<Fixed, Error> {
-    Fixed::random(params)
-}
-
-/// Get the natural logarithm of a fixed-point number.
-///
-/// @param x - The value to calculate the natural logarithm of.
-/// scaled raw value.
-#[wasm_bindgen(skip_jsdoc)]
-pub fn ln(x: Numberish) -> Result<Fixed, Error> {
-    let int = x.to_fixed()?.to_i256()?;
-    let int_result = FixedPoint::ln(int).to_result()?;
-    let result = Fixed {
-        inner: FixedPoint::try_from(int_result).to_result()?,
-        decimals: 18,
-    };
-    Ok(result)
-}
-
 // Types //
-
-#[ts]
-struct GenerateRandomParams {
-    /// The minimum value to generate. Defaults to `0`.
-    min: Option<Numberish>,
-    /// The maximum value to generate. Defaults to 1.0 (scaled) more than `min`.
-    max: Option<Numberish>,
-    /// The number of decimal places to use. Max is `18`. Defaults to `18`.
-    decimals: Option<u8>,
-}
 
 #[wasm_bindgen(typescript_custom_section)]
 const _: &'static str = r#"
@@ -382,4 +311,14 @@ impl ToFixedPoint for Numberish {
             FixedPoint::from_str(&str).to_result()
         }
     }
+}
+
+#[ts]
+struct GenerateRandomParams {
+    /// The minimum value to generate. Defaults to `0`.
+    min: Option<Numberish>,
+    /// The maximum value to generate. Defaults to 1.0 (scaled) more than `min`.
+    max: Option<Numberish>,
+    /// The number of decimal places to use. Max is `18`. Defaults to `18`.
+    decimals: Option<u8>,
 }
